perf(token): narrow a token's type and position to what they hold
There are 34 token types, so Type spent 8 bytes saying one of 34 things and now
spends one. Position held four ints where a document large enough to overflow an
int32 does not fit in memory: 32 bytes down to 16.

The fields are ordered widest first. Left in reading order the three narrow ones
pad out to a word each and the token is 80 bytes; grouped, they share one word
and it is 72.

token.Token is 96 bytes down to 72, and parsing allocates 4.3% fewer bytes for
the same objects. Line, Column, Offset and IndentNum are int32 now, so a caller
doing arithmetic on them against an int converts; comparing them, switching on a
Type and naming a Type constant are unaffected.

+++ encode.go
@@ -522,10 +522,10 @@ func (e *Encoder) encodePtrAnchor(v reflect.Value, column int) ast.Node {
 
 func (e *Encoder) pos(column int) token.Position {
 	return token.Position{
-		Line:      e.line,
-		Column:    column,
-		Offset:    e.offset,
-		IndentNum: e.indentNum,
+		Line:      int32(e.line),
+		Column:    int32(column),
+		Offset:    int32(e.offset),
+		IndentNum: int32(e.indentNum),
 	}
 }
